refactor: remove the mapper in favor of adding the functions to the entity directly to improve dev experience
